refactorizando codigo
