+++ src/css/index.css
@@ -1,0 +1,3 @@
+@import "./auth.css";
+@import "./desktop.css";
+@import "./share.css";
