+++ src/css/share.css
@@ -1,5 +1,3 @@
-@import "./large-shared.css";
-@import "./medium-shared.css";
 @import url('https://fonts.googleapis.com/css2?family=Asap&family=Vollkorn:ital,wght@1,800&display=swap');
 
 #container div {
@@ -63,8 +61,10 @@ display: none;
       border-bottom-width: 2px;
       border-radius: 5px;
     }
+    
     input:focus{
       outline: var(--pink);
       border: 6px var(--pink);
       color: var(--pink);
     }
+
