+++ src/index.html
@@ -6,9 +6,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css">
-  <link rel="stylesheet" href="./modules/index.css">
+  <link rel="stylesheet" href="./css/index.css">
   <link rel="icon" type="image/x-icon" href="https://i.ibb.co/x1gsNXS/pesa-45-rosa.png" />
-  <link href="style.css" rel="stylesheet" />
   <title>GirlFit</title>
 </head>
 
@@ -31,16 +30,16 @@
     </div>
     <div class="view" id="container-sign-up">
       <div class="mobile-logo active"><img src="img/logo/logo-rosa.png"></div>
-        <form>
-          <h2 class="form_title">Antes de Empezar cuentanos un poco sobre ti</h2>
-          <input type="email" id="register-email" placeholder="Correo Electronico">
-          <input type="password" id="register-password" placeholder="Contraseña">
-          <p class="session-text">¿Ya tienes cuenta?<br></p><span id="session-link">Inicia Sesión</span>
-        </form>
-        <button id="btn-sing-up">Registrarse</button>
+      <form>
+        <h2 class="form_title">Antes de Empezar cuentanos un poco sobre ti</h2>
+        <input type="email" id="register-email" placeholder="Correo Electronico">
+        <input type="password" id="register-password" placeholder="Contraseña">
+        <p class="session-text">¿Ya tienes cuenta?<br></p><span id="session-link">Inicia Sesión</span>
+      </form>
+      <button id="btn-sing-up">Registrarse</button>
     </div>
   </div>
-  <script type="module" src="./modules/index.js"></script>
+  <script type="module" src="./js/index.js"></script>
 </body>
 
 </html>
