+++ src/js/index.js
@@ -1,0 +1,7 @@
+import auth from "./indexAuth.js";
+import feed from "./indexFeed.js";
+import shared from "./indexShared.js";
+
+auth();
+feed();
+shared();
