+++ src/js/indexShared.js
