--- src/modules/authenticate/css-authorization/large-authorization.css
@@ -1,11 +1,0 @@
-/* body {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-image: url(img/Imagenes/bruce-mars-ZXq7xoo98b0-unsplash.jpg);
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    backdrop-filter: blur(2px);
-  } */
