--- src/modules/feed/css-feed/index-feed.css
@@ -1,3 +1,0 @@
-@import "./large-feed.css";
-@import "./medium-feed.css";
-
