--- src/modules/feed/javascript-feed/index-feed.js
@@ -1,3 +1,0 @@
-export default function (){
-    console.info("feed initialized");
-};
