--- src/modules/index.css
@@ -1,4 +1,0 @@
-@import "./authenticate/css-authorization/index-authorization.css";
-@import "./feed/css-feed/index-feed.css";
-@import "./shared/css-shared/index-shared.css";
-
