--- src/modules/index.js
@@ -1,7 +1,0 @@
-import authenticate from "./authenticate/javascript-authorization/index-authorization.js";
-import feed from "./feed/javascript-feed/index-feed.js";
-import shared from "./shared/javascript-shared/index-shared.js";
-
-authenticate();
-feed();
-shared();
