--- src/style.css
@@ -1,46 +1,0 @@
-
-
-/* .mobile-logo{
-    width: 20%;
-    display: initial;
-    margin-top: 15%;
-  }*/
-  
-  /* .container-form{
-    display: flex;
-    justify-content: space-around;
-    transition: all .5s ease-out;
-  }
-  
-.formulario {
-    width: 80%;
-    height: 100vh;
-    text-align: center;
-  }
-
-
-
-  .form_title {
-    font-size: 1rem;
-    padding: 1rem 0;
-    margin: 5%;
-  } */
-
- /*  input {
-    height: 7%;
-    width: 100%;
-    display: block;
-    margin: auto;
-    margin-bottom: 1rem;
-    text-align: center;
-    padding: .2rem 0;
-    font-size: .9rem;
-    color: black;
-    background: white;
-    box-shadow: 0 5px 10px -5px;
-    border: none;
-  }*/
-
-
-
-
